+++ cosmos_rl/policy/model/dino_cradio_v3/README.md
@@ -1,0 +1,6 @@
+## Compile the custom kernels
+DeformableAttention module relies on custom CUDA kernels for both increased performance and reduced memory usage. To compile the kernels, run the following command:
+
+```
+python model/deformable_detr/model/ops/setup.py
+```
